+++ examples/with-assistant-transport/app/MyRuntimeProvider.tsx
@@ -2,7 +2,6 @@
 
 import {
   AssistantRuntimeProvider,
-  defineToolkit,
   Tools,
   type AssistantTransportConnectionMetadata,
   unstable_createMessageConverter as createMessageConverter,
@@ -14,57 +13,7 @@ import {
   type LangChainMessage,
 } from "@assistant-ui/react-langgraph";
 import type { ReactNode } from "react";
-import { z } from "zod";
-
-const toolkit = defineToolkit({
-  get_weather: {
-    type: "frontend",
-    description: "Get the current weather for a city",
-    parameters: z.object({
-      location: z.string().describe("The city to get weather for"),
-      unit: z
-        .enum(["celsius", "fahrenheit"])
-        .optional()
-        .describe("Temperature unit"),
-    }),
-    execute: async ({ location, unit = "celsius" }) => {
-      console.log(`Getting weather for ${location} in ${unit}`);
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-
-      const temp = Math.floor(Math.random() * 30) + 10;
-      const conditions = ["sunny", "cloudy", "rainy", "partly cloudy"];
-      const condition =
-        conditions[Math.floor(Math.random() * conditions.length)];
-
-      return {
-        location,
-        temperature: temp,
-        unit,
-        condition,
-        humidity: Math.floor(Math.random() * 40) + 40,
-        windSpeed: Math.floor(Math.random() * 20) + 5,
-      };
-    },
-    streamCall: async (reader) => {
-      console.log("streamCall", reader);
-      const city = await reader.args.get("location");
-      console.log("location", city);
-
-      const args = await reader.args.get();
-      console.log("args", args);
-
-      const result = await reader.response.get();
-      console.log("result", result);
-    },
-    renderText: {
-      running: ({ args }) => `Getting weather for ${args.location ?? "..."}`,
-      complete: ({ args, result }) =>
-        result
-          ? `${args.location}: ${result.temperature}° ${result.unit}, ${result.condition}`
-          : "Weather lookup complete",
-    },
-  },
-});
+import toolkit from "./toolkit";
 
 type MyRuntimeProviderProps = {
   children: ReactNode;
@@ -122,6 +71,10 @@ export function MyRuntimeProvider({ children }: MyRuntimeProviderProps) {
     }),
     body: {
       "Test-Body": "test-value",
+    },
+    prepareSendCommandsRequest: (body) => {
+      console.log("Assistant transport request tools:", body.tools);
+      return body;
     },
     onResponse: () => {
       console.log("Response received from server");
